hp change and make sure we don't run reset_run on gimli

+++ reset_run.sh
@@ -1,5 +1,10 @@
 #!/bin/bash
 
+if [ "$USER" = "gimli" ]; then
+  echo "You are Gimli. Exiting the script."
+  exit 1
+fi
+
 
 rm ./db/db.db 
 rm ./models/*/*.pt
